requirements.txt
fastapi
torchvision
uvicorn
pillow
aiohttp
more_itertools
torch-model-archiver
torchserve
nvgpu
captum

#conda install pytorch torchvision torchaudio cudatoolkit=11.3 -c pytorch


